added all files of the project v1.0
first commit, added all files as they are, maybe some files are redundant.

still need to make changes to the settings menu text (change english and hebrew shortcut to "Inside Math" and "Outside Math" and add a short explanation on how to set the keyboard shortcuts in the readme file.

also i need to test the python server, maybe it's better to create the entire python project in a virtual environment so others wouldn't need to mess around with downloading python and installing packages and setting the path to the python exe and so forth..